Studios 16 and 17 done

+++ Studio16/Studio16Zachary1/TextFile.cpp
@@ -7,13 +7,12 @@ TextFile::TextFile(string myName): fileName(myName){
 }
 
 void TextFile::read(){
-	for (unsigned int i = 0; i < fileContents.size(); i++) {
+	for (int i = 0; i < fileContents.size(); i++) {
 		cout << fileContents.at(i);
 	}
 }
 
 int TextFile::write(vector<char> myVec){
-	// Does this successfully update the contents?
 	fileContents = myVec;
 	// Update this to return 0 if successful
 	return success;
@@ -22,7 +21,7 @@ int TextFile::write(vector<char> myVec){
 int TextFile::append(vector<char> myVec)
 {
 	// Append the contents of myVece to the end of FileContents
-	for (unsigned int i = 0; i < myVec.size(); i++) {
+	for (int i = 0; i < myVec.size(); i++) {
 		fileContents.push_back(myVec.at(i));
 	}
 	return success;
@@ -30,7 +29,7 @@ int TextFile::append(vector<char> myVec)
 
 unsigned int TextFile::getSize()
 {
-	return fileContents.size();
+	return (unsigned int)fileContents.size();
 }
 
 string TextFile::getName()
